+++ convert_str.c
@@ -7,7 +7,7 @@ char	*convert_str(va_list ap, t_pformat *pformat)
 
 	str = str == NULL ? ft_strdup("(null)") : ft_strdup(str);
 
-	if (pformat->precision != -1 && pformat->precision < ft_strlen(str))
+	if (pformat->precision != -1 && pformat->precision < (int)ft_strlen(str))
 		str[pformat->precision] = '\0';
 	str = handle_padding(pformat, str);
